requirements_linux.txt
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
DateTime==4.2
docutils==0.14
html5lib==1.0.1
idna==2.7
parse==1.8.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
tzlocal==1.5.1
urllib3==1.23
webencodings==0.5.1
zope.interface==4.5.0
